requirements.txt
# python3.8.18  CUDA10.2

# pip install -r requirements.txt

# base ----------------------------------------

matplotlib>=3.4.3
numpy>=1.21.6
opencv-python>=4.5.3.56
Pillow>=9.2.0
PyYAML>=5.4.1
scipy>=1.7.1
# torch>=1.7.1
# torchvision>=0.8.2
tqdm>=4.64.1

# plotting ------------------------------------

seaborn>=0.11.2
pandas>=1.3.2

# deep_sort -----------------------------------

easydict>=1.9

requests>=2.26.0
